requirements.txt
fastapi
uvicorn
python-dotenv
supabase
razorpay
firebase-admin
pydantic